Skip parsed_uri function
`parsed_uri` uses a regex to figure out the (scheme|host|port|path).
This breaks on the uri's we're using `https://subdomain2.subdomain1.domain/path1/path2/1`

Skip the function call and regex and pass in a table with that
information to httpc:request_uri(). The proper request body will be
returned.